rewrite AddressInput
