+++ src/components/inputs/AddressInput.js
@@ -1,5 +1,4 @@
 import React, { useState, useEffect } from 'react';
-import axios from 'axios';
 import { connect } from 'react-redux';
 import { withRouter } from 'react-router-dom';
 import { getValuation } from '../../actions/landingpageActions';
@@ -7,57 +6,39 @@ import { useWindowWidth } from '../../helper-functions/display-functions';
 import Loader from 'react-loader-spinner';
 
 function AddressInput({ history, getValuation, fetching }) {
-  const [sessiontoken, setSessiontoken] = useState('');
+  const [sessionToken, setSessionToken] = useState('');
   const [address, setAddress] = useState('');
   const [predictions, setPredictions] = useState([]);
 
   // Variables for Google Place API
-  const location = '43.3148,-85.6024'; // Latitude, longtitude of Michigan
-  const types = 'address';
-  const key = 'AIzaSyBQG-Y3BtowkEvTBq3dPPROa-GuMm1Rfpk';
+  const google = window.google;
+  const location = new google.maps.LatLng(43.3148, -85.6024); // Latitude, longtitude of Michigan
 
   // Get session token on first render
   useEffect(() => {
-    axios
-      .get('https://www.uuidgenerator.net/api/version4')
-      .then(res => {
-        setSessiontoken(res.data);
-      })
-      .catch(err => {
-        console.log(err);
-      });
+    setSessionToken(new google.maps.places.AutocompleteSessionToken());
   }, []);
+
+  const displaySuggestions = (predictions, status) => {
+    if (status !== google.maps.places.PlacesServiceStatus.OK) {
+      alert(status);
+      return;
+    }
+    setPredictions(predictions);
+  };
 
   const handleInputChange = e => {
     const string = e.target.value;
     setAddress(string);
-
-    // Autocomplete request is called every few keystrokes
+    var service = new google.maps.places.AutocompleteService();
     if ([3, 6, 10, 15].includes(string.length)) {
-      const url = `https://cors-anywhere.herokuapp.com/https://maps.googleapis.com/maps/api/place/autocomplete/json?input=${string}&types=${types}&location=${location}&radius=500&key=${key}&sessiontoken=${sessiontoken}`;
-      axios
-        .get(url)
-        .then(res => {
-          setPredictions(res.data.predictions);
-        })
-        .catch(err => {
-          console.log(err);
-        });
+      service.getQueryPredictions({ input: string, sessionToken: sessionToken, location: location, radius: 500, types: ['address'] }, displaySuggestions);
     }
   };
 
   const getValue = e => {
     e.preventDefault();
     getValuation(address, history);
-    // axios
-    //   .post('https://labs12-real-estate.herokuapp.com/api/houses/getvalue', {
-    //     address
-    //   })
-    //   .then(state => {
-    //     localStorage.setItem('initialData', JSON.stringify(state));
-    //     history.push('/wizard-form');
-    //   })
-    //   .catch(console.error);
   };
 
   const fillAddress = (address, e) => {
@@ -80,7 +61,7 @@ function AddressInput({ history, getValuation, fetching }) {
           )}
         </form>
         <div className="search_result_container">
-          {predictions.length > 0 && address ? (
+          {predictions.length > 2 && address ? (
             <div className="search_result_dropdown">
               {predictions.map(prediction => (
                 <button key={prediction.id} onClick={e => fillAddress(prediction.description, e)}>
